renaming file and includes to RTES standard style
