--- application/include/foo.h
@@ -1,9 +1,0 @@
-
-class Foo {
-public:
-    int do_something(char);
-    Foo(int);
-
-private:
-    int n;
-};
