+++ application/include/midi-scaler.hpp
