+++ application/include/patterns.hpp
@@ -3,7 +3,7 @@
 #include <functional>
 #include <atomic>
 #include <thread>
-#include "ILedDriver.hpp"
+#include "i-led-driver.hpp"
 
 namespace led_pattern {
