+++ application/include/voltage-scaler.hpp
