+++ application/src/chord-manager.cpp
@@ -1,4 +1,4 @@
-#include "ChordManager.hpp"
+#include "chord-manager.hpp"
 
 
 void ChordManager::updateChord(int index){
