+++ application/src/main.cpp
@@ -1,9 +1,9 @@
-#include "SynthController.hpp"
-#include "TLC59711.h"
-#include "adc-driver.h"
-#include "button-driver.h"
-#include "flex-sensor.h"
-#include "gpio.h"
+#include "synth-controller.hpp"
+#include "tlc59711.hpp"
+#include "adc-driver.hpp"
+#include "button-driver.hpp"
+#include "flex-sensor.hpp"
+#include "gpio.hpp"
 #include "midi-driver.hpp"
 #include "patterns.h"
 #include "voltage-scaler.h"
