--- application/src/test_leds_buttons.cpp
@@ -1,86 +1,0 @@
-#include "SynthController.hpp"
-#include "TLC59711.h"
-#include <array>
-#include <chrono>
-#include <cmath>
-#include <cstdint>
-#include <cstdio>
-#include <cstring>
-#include <fcntl.h>
-#include <gpiod.hpp>
-#include <iostream>
-#include "button-driver.h"
-#include "gpio.h"
-#include "patterns.h"
-#include <linux/spi/spi.h>
-#include <linux/spi/spidev.h>
-#include <sys/ioctl.h>
-#include <thread>
-#include <unistd.h>
-#include <ads1115rpi.h>
-
-#include <chrono>
-#include <csignal>
-#include <iostream>
-#include <thread>
-
-using namespace std::chrono_literals;
-
-static bool g_running = true;
-
-void signalHandler(int) {
-    g_running = false;
-}
-
-int main() {
-    std::signal(SIGINT, signalHandler);
-
-    // Replace these with your real GPIO pins / driver count
-    constexpr int DATA_PIN = 17;
-    constexpr int CLK_PIN  = 27;
-
-    try {
-        gpio::setupGpio();
-        TLC59711 tlc(DATA_PIN, CLK_PIN);
-        tlc.start();
-
-        PatternFade fade(tlc);
-        
-        SynthController synth(tlc);
-
-        std::cout << "Starting LED integration test...\n";
-
-        // Give the TLC worker a moment to start
-        std::this_thread::sleep_for(500ms);
-
-        // Assumed button mapping:
-        // 0 -> EQ
-        // 1 -> SOURCE_EQ
-        // 2 -> DETUNE
-        // 3 -> CHORD
-        //
-        // Adjust these if your ModeManager uses different indices.
-        
-        ButtonDriver bd;
-        
-        bd.registerSingleButtonCallback([&synth] (int index) {
-            std::cout << index << std::endl;
-            synth.onButtonEvent(index);
-        });
-        
-        bd.registerAllButtonsCallback([&synth] () {
-            std::cout << "all buttons\n";
-            synth.onButtonEvent(2);
-        });
-    
-        getchar();
-        
-        tlc.stop();
-    }
-    catch (const std::exception& e) {
-        std::cerr << "Test failed: " << e.what() << '\n';
-        return 1;
-    }
-
-    return 0;
-}
